requirements.txt
Django==1.11.4
django-ckeditor==5.4.0
django-js-asset==1.1.0
django-taggit==0.22.2
Pillow==5.1.0
PyMySQL==0.8.0
pytz==2018.4
